pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.12.1,<2.0.0",
    "fastapi[standard]>=0.114.2,<1.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "psycopg[binary]>=3.1.13,<4.0.0",
    "pydantic-settings>=2.2.1,<3.0.0",
    "sqlmodel>=0.0.21,<1.0.0",
    "tenacity>=8.2.3,<9.0.0",
]

[dependency-groups]
dev = [
    "coverage>=7.4.3,<8.0.0",
    "pre-commit>=3.6.2,<4.0.0",
    "pytest>=7.4.3,<8.0.0",
    "ruff>=0.2.2,<1.0.0",
]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
